Added option to paste URLs
